Obtain the dst address when reading UDP at the dispatcher.
The destination address can be used to de-multiplex the calls made
to the dispatcher which should go to the different colibri services
instead.
The actual logic of getting the destination IP address is missing,
and only changes to the signatures and types are effected in this commit.
