+++ go/dispatcher/internal/respool/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//go/lib/common:go_default_library",
         "//go/lib/serrors:go_default_library",
         "//go/lib/slayers:go_default_library",
+        "//go/lib/underlay/conn:go_default_library",
         "@com_github_google_gopacket//:go_default_library",
     ],
 )
